+++ Models/ViewModels/ReservationViewModel.cs
@@ -5,8 +5,8 @@ namespace RVBARBER.ViewModels;
 public class ReservationViewModel
 {
     public List<string> Trattamenti { get; set; } = [];
-    public decimal Totale { get; set; }
-    public decimal Durata { get; set; }
+    public double Totale { get; set; }
+    public double Durata { get; set; }
     public DateTime Slot { get; set; }
     public string Barbiere { get; set; } = null!;
     public string Cliente { get; set; } = null!;
