+++ Models/ViewModels/TreatmentViewModel.cs
@@ -4,7 +4,7 @@ public class TreatmentViewModel
 {
     public string NomeTrattamento { get; set; } = null!;
     public string? Descrizione { get; set; }
-    public decimal Prezzo { get; set; }
+    public double Prezzo { get; set; }
     public int DurataMinuti { get; set; }
     public string Categoria { get; set; } = null!;
 }
